DRAFT: Add support for saving full document
Most of the code is similar to the export functionality. The exception
is that this document will encompass additional non-svg data.

The `SerializeData` request accepts an `action` string that it opaquely
passes through. The response handler can use the `action` string to
decide if the serialization is for saving or auto-saving.