Update MIME type to reflect downloaded file
When a file is exported to SVG, the existing code sets
the SVG mime type. When we trying to download the document
as a custom "gsvg" format, the existing code was force
saving the file as an "svg" due to the MIME type.

So in this commit, we change the code to use regular text
MIME type for non-svg files.